requirements.txt
pandas
openpyxl
PyPDF2
reportlab
requests
python-dotenv
